Release cms-item 2.4.1:  * sizeFound now returns long instead of int. There's potentially a huge amount of hits for a query.

+++ src/main/java/se/simonsoft/cms/item/list/CmsItemList.java
@@ -38,7 +38,7 @@ public interface CmsItemList extends Iterable<CmsItem> {
 	 * Based on the query that produced the list.
 	 * @return the total number of items matching the iteration criteria
 	 */	
-	public int sizeFound();
+	public long sizeFound();
 	
 	/**
 	 * @return {@link CmsItemListMetaMap}
